refactor: restructure module imports and update documentation paths
