+++ pyproject.toml
@@ -29,7 +29,7 @@ requires = ["hatchling"]         # uv works with any PEP-517 builder
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["sparkdantic"]
+packages = ["src/sparkdantic"]
 
 [dependency-groups]
 dev = [
